models/DbProfile.py
# models.py
from sqlalchemy import Column, Integer, String
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.types import JSON

Base = declarative_base()


class DbProfile(Base):
    __tablename__ = "profiles"

    id = Column(Integer, primary_key=True, index=True)

    # Basic fields stored in separate columns
    name = Column(String, nullable=False)
    phone = Column(String, nullable=False)
    email = Column(String, nullable=False)
    summary = Column(String)

    # Additional fields stored as JSON (dictionary)
    extra_data = Column(JSON, nullable=True)
